Use os._exit instead of sys.exit
The latter propagates a SystemExit exception in the transient child
forks.  It is unlikely a user wants exception unwinding to happen
several times when daemonizing, rather, just once when the process ends.

Finally, this prevents drop-in compatibility with the older
python-daemon package.  It exits processes this way, and the difference
does amount to rendering pep3143daemon unusable with py.test, which
emits very confused output when SystemExit is raised by sys.exit several
times.